requirements.txt
flask==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0
google-auth==2.16.2
google-cloud-storage==2.7.0
google-generativeai==0.3.1
requests==2.28.2